+++ serena_shop/src/main/java/com/serena/serena_shop/model/DetalleVenta.java
@@ -24,7 +24,7 @@ public class DetalleVenta {
     private Integer cantidad;
 
     @Column(name = "precio_unitario")
-    private Integer precioUnitario;
+    private Double precioUnitario;
 
     @Column(name = "subtotal", insertable=false, updatable=false)
     private Integer subtotal;
